AI/rag_engine/requirements.txt
# HumanLoop RAG Engine - Local Dependencies
# No external APIs. Everything runs locally.

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector store
faiss-cpu>=1.7.4

# HTTP client for Ollama
requests>=2.28.0

# NumPy (usually pulled by sentence-transformers/faiss)
numpy>=1.24.0
